feat: update PWA

+++ pwa/components/book/Show.tsx
@@ -1,45 +1,44 @@
-import { FunctionComponent, useState } from 'react';
-import Link from 'next/link';
-import Image from 'next/image';
-import { useRouter } from 'next/router';
-import { fetch } from 'utils/dataAccess';
-import { Book } from 'types/Book';
+import { FunctionComponent, useState } from "react";
+import Link from "next/link";
+import { useRouter } from "next/router";
+import Head from "next/head";
+
+import ReferenceLinks from "../common/ReferenceLinks";
+import { fetch, getPath } from "../../utils/dataAccess";
+import { Book } from "../../types/Book";
 
 interface Props {
   book: Book;
+  text: string;
 }
 
-export const Show: FunctionComponent<Props> = ({ book }) => {
-  const [error, setError] = useState(null);
+export const Show: FunctionComponent<Props> = ({ book, text }) => {
+  const [error, setError] = useState<string | null>(null);
   const router = useRouter();
 
   const handleDelete = async () => {
-    if (!window.confirm('Are you sure you want to delete this item?')) return;
+    if (!book["@id"]) return;
+    if (!window.confirm("Are you sure you want to delete this item?")) return;
 
     try {
-      await fetch(book['@id'], { method: 'DELETE' });
-      router.push('/books');
+      await fetch(book["@id"], { method: "DELETE" });
+      router.push("/books");
     } catch (error) {
-      setError('Error when deleting the resource.');
-      console.error(error);
-    }
-  };
-
-  const handleGenerateCover = async () => {
-    try {
-      await fetch(`${book['@id']}/generate-cover`, {
-        method: 'PUT',
-        body: JSON.stringify({}),
-      });
-    } catch (error) {
-      setError('Error when generating the cover.');
+      setError("Error when deleting the resource.");
       console.error(error);
     }
   };
 
   return (
     <div>
-      <h1>{`Show Book ${book['@id']}`}</h1>
+      <Head>
+        <title>{`Show Book ${book["@id"]}`}</title>
+        <script
+          type="application/ld+json"
+          dangerouslySetInnerHTML={{ __html: text }}
+        />
+      </Head>
+      <h1>{`Show Book ${book["@id"]}`}</h1>
       <table className="table table-responsive table-striped table-hover">
         <thead>
           <tr>
@@ -50,54 +49,33 @@ export const Show: FunctionComponent<Props> = ({ book }) => {
         <tbody>
           <tr>
             <th scope="row">isbn</th>
-            <td>{book['isbn']}</td>
+            <td>{book["isbn"]}</td>
           </tr>
           <tr>
             <th scope="row">title</th>
-            <td>{book['title']}</td>
+            <td>{book["title"]}</td>
           </tr>
           <tr>
             <th scope="row">description</th>
-            <td>{book['description']}</td>
+            <td>{book["description"]}</td>
           </tr>
           <tr>
             <th scope="row">author</th>
-            <td>{book['author']}</td>
+            <td>{book["author"]}</td>
           </tr>
           <tr>
             <th scope="row">publicationDate</th>
-            <td>{book['publicationDate']}</td>
+            <td>{book["publicationDate"]?.toLocaleString()}</td>
           </tr>
           <tr>
             <th scope="row">reviews</th>
             <td>
-              {book['reviews'] &&
-                book['reviews'].length !== 0 &&
-                book['reviews'].map((review: string) => (
-                  <div key={review}>
-                    <Link href={review}>{review}</Link>
-                  </div>
-                ))}
-            </td>
-          </tr>
-          <tr>
-            <th scope="row">cover</th>
-            <td>
-              {(book['cover'] && (
-                <Image
-                  alt="Book cover"
-                  src={book['cover']}
-                  width={500}
-                  height={500}
-                />
-              )) || (
-                <button
-                  id="generate-cover"
-                  className="btn btn-primary"
-                  onClick={handleGenerateCover}>
-                  <a>Generate cover</a>
-                </button>
-              )}
+              <ReferenceLinks
+                items={book["reviews"].map((emb: any) => ({
+                  href: getPath(emb["@id"], "/reviews/[id]"),
+                  name: emb["@id"],
+                }))}
+              />
             </td>
           </tr>
         </tbody>
@@ -109,12 +87,12 @@ export const Show: FunctionComponent<Props> = ({ book }) => {
       )}
       <Link href="/books">
         <a className="btn btn-primary">Back to list</a>
-      </Link>{' '}
-      <Link href={`${book['@id']}/edit`}>
+      </Link>{" "}
+      <Link href={getPath(book["@id"], "/books/[id]/edit")}>
         <a className="btn btn-warning">Edit</a>
       </Link>
       <button className="btn btn-danger" onClick={handleDelete}>
-        <a>Delete</a>
+        Delete
       </button>
     </div>
   );
